Rework Wallet.sendMessage signature to work w/ limitations on Kotlin generics, move logging/low-level HTTP fake-ipc wire into libkylons, write minimal devdevwallet binding, implement getTrades ipc message.
