+++ libkylons/src/main/kotlin/com/pylons/lib/Wallet.kt
@@ -1,6 +1,6 @@
 package com.pylons.lib
 
-import com.beust.klaxon.Klaxon
+import com.pylons.ipc.HttpIpcWire
 import com.pylons.ipc.Message
 import com.pylons.lib.types.*
 import com.pylons.lib.types.tx.Coin
@@ -10,20 +10,17 @@ import com.pylons.lib.types.tx.recipe.EntriesList
 import com.pylons.lib.types.tx.recipe.ItemInput
 import com.pylons.lib.types.tx.recipe.WeightedOutput
 import com.pylons.lib.types.tx.Trade
+import kotlin.reflect.KClass
 
 /**
  * Generic high-level interface between JVM clients and a Pylons wallet.
- * TODO: port over the actual tx type so we can resolve messages that
- * return a tx (expect this to be messy)
  */
 abstract class Wallet {
-    val klaxon = Klaxon()
-
     /**
      * Signature for the method what we call to pass messages into the
      * wallet. IPC happens after this implementation.
      */
-    protected abstract fun <T> sendMessage(message: Message, callback : (T) -> Unit)
+    protected abstract fun sendMessage(outType : KClass<*>, message: Message, callback : (Any?) -> Unit)
 
     /**
      * True if an IPC target exists; false otherwise.
@@ -34,50 +31,56 @@ abstract class Wallet {
      * Get current profile, or null if none exists.
      */
     fun fetchProfile (callback: (Profile?) -> Unit) {
-        sendMessage<Profile?>(Message.GetProfile()) {callback(it)}
+        sendMessage(Profile::class, Message.GetProfile()) {callback(it as Profile?)}
     }
 
     /**
      * Get a list of all items owned by current profile.
      */
     fun listItems(callback: (List<Item>) -> Unit) {
-        sendMessage<Profile?>(Message.GetProfile()) {callback(it?.items.orEmpty())}
+        sendMessage(Profile::class, Message.GetProfile()) {callback((it as Profile?)?.items.orEmpty())}
     }
 
     /**
      * Register a new profile.
      */
     fun registerProfile (callback: (Profile?) -> Unit) {
-        sendMessage<Profile?>(Message.RegisterProfile()) {callback(it)}
+        sendMessage(Profile::class, Message.RegisterProfile()) {callback(it as Profile?)}
     }
 
     fun placeForSale (item : Item, price : Long, callback: (Transaction?) -> Unit) {
-        sendMessage<Transaction?>(Message.CreateTrade(listOf(
+        sendMessage(Transaction::class, Message.CreateTrade(listOf(
             klaxon.toJsonString(Coin("pylon", price))),
-            listOf(), listOf(), listOf(item.id))) {callback(it)}
+            listOf(), listOf(), listOf(item.id))) {callback(it as Transaction?)}
     }
 
     fun getTrades(callback: (List<Trade>?) -> Unit) {
-        // the message to resolve this doesn't exist in walletcore rn! whoops!
-        // let's get that before we get ipc online ig
-        throw NotImplementedError()
+        sendMessage(List::class, Message.GetTrades()) {callback(it as List<Trade>?)}
     }
 
     fun buyItem (trade : Trade, callback: (Transaction?) -> Unit) {
-        sendMessage<Transaction?>(Message.FulfillTrade(trade.id)) {callback(it)}
+        sendMessage(Transaction::class, Message.FulfillTrade(trade.id)) {callback(it as Transaction?)}
     }
 
     fun createRecipe(name : String, cookbook : String, description : String,
                      blockInterval : Long, coinInputs : List<CoinInput>,
                      itemInputs: List<ItemInput>, outputTable : EntriesList,
                      outputs : List<WeightedOutput>, callback: (Transaction?) -> Unit) {
-        sendMessage<Transaction>(Message.CreateRecipes(listOf(name), listOf(cookbook), listOf(description),
+        sendMessage(Transaction::class, Message.CreateRecipes(listOf(name), listOf(cookbook), listOf(description),
         listOf(blockInterval), listOf(klaxon.toJsonString(coinInputs)), listOf(klaxon.toJsonString(itemInputs)),
-        listOf(klaxon.toJsonString(outputTable)), listOf(klaxon.toJsonString(outputs)))) {callback(it)}
+        listOf(klaxon.toJsonString(outputTable)), listOf(klaxon.toJsonString(outputs)))) {callback(it as Transaction?)}
     }
 
-    class Android : Wallet(){
-        override fun <T> sendMessage(message: Message, callback: (T) -> Unit) {
+    fun android() : AndroidWallet = AndroidWallet.instance
+
+    fun devDevWallet() : DevDevWallet = DevDevWallet.instance
+
+    class AndroidWallet : Wallet(){
+        companion object {
+            val instance : AndroidWallet = AndroidWallet()
+        }
+
+        override fun sendMessage(outType : KClass<*>, message: Message, callback: (Any?) -> Unit) {
             TODO("Not yet implemented")
         }
 
@@ -86,16 +89,22 @@ abstract class Wallet {
         }
     }
 
-    // todo: idk when we'll have android wallet online so
-    // once i like this library i should throw together this
-    // binding and do a quick and nasty client for it as a sanity check
     class DevDevWallet : Wallet() {
-        override fun <T> sendMessage(message: Message, callback: (T) -> Unit) {
-            TODO("Not yet implemented")
+        companion object {
+            val instance : DevDevWallet = DevDevWallet()
+        }
+
+        override fun sendMessage(outType : KClass<*>, message: Message, callback: (Any?) -> Unit) {
+            /// HttpIpcWire is dead simple; it just writes a string.
+            HttpIpcWire.writeString(klaxon.toJsonString(message))
+            // B/c HttpIpcWire is extremely simple, calling readMessage means we
+            // just wait forever until we get something. That's fine - you just
+            // want to use a worker thread to interact w/ your Wallet instances.
+            callback(klaxon.parser(outType).parse(HttpIpcWire.readMessage().orEmpty()))
         }
 
         override fun exists(callback: (Boolean) -> Unit) {
-            TODO("Not yet implemented")
+            callback(true) // todo: devdevwallet doesn't handle connection breaking yet so we can't tell if it's connected.
         }
     }
 }
