+++ libkylons/src/main/kotlin/com/pylons/lib/logging/LogTag.kt
@@ -1,4 +1,4 @@
-package com.pylons.wallet.core.logging
+package com.pylons.lib.logging
 
 object LogTag {
     const val info = "INFO"
